Rename smoke tests to integration tests
Resolves #25
