+++ pom.xml
@@ -242,7 +242,7 @@
                         <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-surefire-plugin</artifactId>
                         <configuration>
-                            <excludedGroups>smoke</excludedGroups>
+                            <excludedGroups>integration-test</excludedGroups>
                         </configuration>
                     </plugin>
                 </plugins>
